# docs/swagger.yaml
basePath: /api
definitions:
...
  tokenhandler.OidcTokenRequest:
    properties:
      idp_token:
        type: string
    type: object
  tokenhandler.TokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  tokenhandler.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.MessageResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact:
    email: denglerchr@gmail.com
    name: Christian Dengler
    url: https://www.swcd.lu
  description: This is a server for a image, audio and file storage with integrated
    web-ui.
  title: SWCD MediaHub-API
  version: 2.0.0-beta1
paths:
  /api/logout:
    post:
      consumes:
      - application/json
      description: Revokes a Refresh Token, effectively logging the user out.
      parameters:
      - description: Refresh token JSON body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tokenhandler.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully message
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - token
  /api/token:
    post:
      consumes:
      - application/json
      description: |-
        Obtains an internal JWT Access/Refresh token pair.
        Supports two authentication methods:
        1. Local Authentication: Send standard Basic Auth headers.
        2. OIDC Token Exchange (commercial version only): Send a JSON body containing a valid external JWT (`idp_token`).
        Providing both methods in a single request will result in a 400 Bad Request.
      parameters:
      - description: OIDC Identity Provider Token (required if not using Basic Auth)
        in: body
        name: body
        schema:
          $ref: '#/definitions/tokenhandler.OidcTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns access and refresh tokens
          schema:
            $ref: '#/definitions/tokenhandler.TokenResponse'
        "400":
          description: Ambiguous authentication request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Invalid credentials, invalid OIDC token, or missing authentication
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error or OIDC not available
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get a token pair
      tags:
      - token
  /api/token/refresh:
    post:
      consumes:
      - application/json
      description: Uses a valid Refresh Token to obtain a new Access/Refresh token
        pair.
      parameters:
      - description: Refresh token JSON body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tokenhandler.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns new access and refresh tokens
          schema:
            $ref: '#/definitions/tokenhandler.TokenResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Refresh token pair
      tags:
      - token
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"